requirements.txt
# Core dependencies
google-genai>=0.3.0
streamlit>=1.29.0
python-dotenv>=1.0.0

# Memory dependencies
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Utils
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0

# Environment
python-dotenv>=1.0.0
